feat:ui update des scenes tout est clean mtn
