--- Scenes/Guillaume/controls_menu.tscn
@@ -1,50 +1,0 @@
-[gd_scene format=3 uid="uid://d1wq5r3s0w6dh"]
-
-[node name="Controls_Menu" type="Control"]
-visible = false
-layout_mode = 3
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 1
-anchors_preset = 8
-anchor_left = 0.5
-anchor_top = 0.5
-anchor_right = 0.5
-anchor_bottom = 0.5
-offset_left = -556.0
-offset_top = -304.0
-offset_right = 556.0
-offset_bottom = 304.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="Mouse Settings" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Mouse Settings"
-
-[node name="Keyboard Settings" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Keyboard Settings"
-
-[node name="MarginContainer2" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 551.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer2"]
-layout_mode = 2
-
-[node name="Exit" type="Button" parent="MarginContainer2/VBoxContainer"]
-layout_mode = 2
-text = "
-Exit
-
-"
