--- Scenes/Guillaume/credits_menu.tscn
@@ -1,59 +1,0 @@
-[gd_scene format=3 uid="uid://c6ecyb70d4tkl"]
-
-[node name="Credits_Menu" type="Control"]
-visible = false
-layout_mode = 3
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 20.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="Matteo Dupond" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Matteo Dupond
-"
-
-[node name="Lilian Labat" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Lilian Labat
-"
-
-[node name="Guillaume Dufour" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Guillaume Dufour"
-
-[node name="Nelson Rudent" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Nelson Rudent
-"
-
-[node name="Luan Ray-Chagas" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Luan Ray-Chagas
-"
-
-[node name="MarginContainer2" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 551.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer2"]
-layout_mode = 2
-
-[node name="Exit" type="Button" parent="MarginContainer2/VBoxContainer"]
-layout_mode = 2
-text = "
-Exit
-
-"
