--- Scenes/Guillaume/graphics_menu.tscn
@@ -1,48 +1,0 @@
-[gd_scene format=3 uid="uid://dpo0dpr1yu5lv"]
-
-[node name="Graphics_Menu" type="Control"]
-visible = false
-layout_mode = 3
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 1
-anchors_preset = 8
-anchor_left = 0.5
-anchor_top = 0.5
-anchor_right = 0.5
-anchor_bottom = 0.5
-offset_left = -276.0
-offset_top = -224.0
-offset_right = 276.0
-offset_bottom = 224.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="HSlider" type="HSlider" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-size_flags_vertical = 6
-step = 10.0
-value = 100.0
-
-[node name="MarginContainer2" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 551.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer2"]
-layout_mode = 2
-
-[node name="Exit" type="Button" parent="MarginContainer2/VBoxContainer"]
-layout_mode = 2
-text = "
-Exit
-
-"
