--- Scenes/Guillaume/main_menu.tscn
@@ -1,45 +1,0 @@
-[gd_scene format=3 uid="uid://dww6x4dpar6um"]
-
-[node name="Main_Menu" type="Control"]
-layout_mode = 3
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 19.0
-offset_right = 1132.0
-offset_bottom = 627.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="Play" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Play"
-
-[node name="Credits" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Credits"
-
-[node name="Options" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Options"
-
-[node name="MarginContainer2" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 1043.0
-offset_top = 597.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer2"]
-layout_mode = 2
-
-[node name="Quit" type="Button" parent="MarginContainer2/VBoxContainer"]
-layout_mode = 2
-text = "Quit"
-
-[connection signal="pressed" from="MarginContainer/VBoxContainer/Options" to="." method="_on_options_pressed"]
