--- Scenes/Guillaume/options_menu.tscn
@@ -1,47 +1,0 @@
-[gd_scene format=3 uid="uid://cnysnereo0pqb"]
-
-[node name="Options_Menu" type="Control"]
-visible = false
-layout_mode = 3
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 20.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="Sound" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Sound"
-
-[node name="Graphics" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Graphics"
-
-[node name="Controls" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Binds"
-
-[node name="MarginContainer2" type="MarginContainer" parent="."]
-layout_mode = 0
-offset_left = 20.0
-offset_top = 551.0
-offset_right = 1132.0
-offset_bottom = 628.0
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer2"]
-layout_mode = 2
-
-[node name="Exit" type="Button" parent="MarginContainer2/VBoxContainer"]
-layout_mode = 2
-text = "
-Exit
-
-"
