+++ project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="Atria"
-run/main_scene="res://Scenes/Guillaume/Root_Node.tscn"
+run/main_scene="res://Scenes/Guillaume/UI.tscn"
 config/features=PackedStringArray("4.2", "C#", "Forward Plus")
 config/icon="res://icon.svg"
